pyproject.toml
[project]
name = "data-task-solution"
version = "0.1.0"
description = "Solution RMT Data Engineering Task (Gerardo Corado)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "psycopg[binary]>=3.2.6",
    "ruff>=0.11.0",
    "apache-airflow",
    "airflow-clickhouse-plugin",
    "clickhouse-connect",
    "apache-airflow-providers-common-sql",
    "apache-airflow-providers-postgres",
    "apache-airflow-providers-common-sql[pandas]",
    "jupyterlab"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
